+++ packages/blinkid-wasm/src/session/BlinkIdScanningResult.ts
@@ -93,6 +93,14 @@ export type BlinkIdScanningResult = {
   specificDocumentValidity?: StringResult;
   /** The visa type of the document */
   visaType?: StringResult;
+  /** The vehicle owner */
+  vehicleOwner?: StringResult;
+  /** The certificate number of the document owner */
+  certificateNumber?: StringResult;
+  /** The country code of the document owner */
+  countryCode?: StringResult;
+  /** The national insurance number of the document owner */
+  nationalInsuranceNumber?: StringResult;
 
   /** The date of birth of the document owner */
   dateOfBirth?: DateResult<StringResult>;