+++ packages/blinkid-wasm/src/settings/CroppedImageSettings.ts
@@ -7,7 +7,10 @@ export type CroppedImageSettings = {
   /** The DPI value for the cropped image. */
   dotsPerInch: number;
 
-  /** The extension factor for the cropped image. Allowed value is from `0.0` to `1.0` */
+  /**
+   * The extension factor for the cropped image. Applicable only to
+   * document images. Allowed value is from `0.0` to `1.0`
+   */
   extensionFactor: number;
 
   /** Indicates whether the document image should be returned. */