+++ apps/examples/blinkid-advanced-setup/CHANGELOG.md
@@ -1,5 +1,14 @@
 # example-blinkid-advanced-setup
 
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/camera-manager@7.2.0
+  - @microblink/blinkid-ux-manager@7.3.0
+  - @microblink/blinkid-core@7.3.0
+
 ## 1.1.6
 
 ### Patch Changes