+++ apps/examples/blinkid-core-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-blinkid-core-api
 
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/blinkid-core@7.3.0
+
 ## 1.1.6
 
 ### Patch Changes